refactor adapter pattern
