+++ Structural/Adapter/src/interfaces/media/AdvancedMediaPlayer.ts
@@ -1,0 +1,4 @@
+interface AdvancedMediaPlayer {
+  playVid(fileName: string): void;
+  playMusic(fileName: string): void;
+}
