+++ Structural/Adapter/src/interfaces/media/MediaPlayer.ts
@@ -1,0 +1,3 @@
+interface MediaPlayer {
+  play(mediaType: string, fileName: string): void;
+}
