+++ Structural/Adapter/src/interfaces/payment/CreditCardPayment.ts
@@ -1,0 +1,4 @@
+export interface CreditCardPayment {
+  getBankDetails(): void;
+  getCreditCard(): string;
+}
