+++ Structural/Adapter/src/models/media/Mp4Player.ts
@@ -1,0 +1,9 @@
+import { AdvancedMediaPlayer } from "../../interfaces/media/AdvancedMediaPlayer";
+
+export class Mp4Player implements AdvancedMediaPlayer {
+  playVid(fileName: string): void {}
+
+  playMusic(fileName: string): void {
+    console.log(`Play Music Player`);
+  }
+}
