+++ Structural/Adapter/src/models/media/YouTubePlayer.ts
@@ -1,0 +1,9 @@
+import { AdvancedMediaPlayer } from "../../interfaces/media/AdvancedMediaPlayer";
+
+export class YouTubePlayer implements AdvancedMediaPlayer {
+  playVid(fileName: string) {
+    console.log(`opened youtube video: ${fileName}`);
+  }
+
+  playMusic(fileName: string): void {}
+}
